modernized release script

+++ release.sh
@@ -1,7 +1,14 @@
 #!/bin/sh
-# automate release testing steps
+# automate release steps
 #
+#   --dry-run  - tell bumpversion to not make changes
 # check local tests pass
+if [ -e /usr/local/bin/gmake ]
+then 
+	MAKE=gmake
+else
+	MAKE=make
+fi
 
 if hg sum --remote
     then echo "** no remote changegs **"
@@ -9,8 +16,9 @@ if hg sum --remote
     echo "** FAIL ** repository out of sync"
     exit 1
 fi
+
 # check setup.py required setup.py attributes
-if python setup.py check --strict
+if python setup.py check -s
     then echo "** Attribute Testing passed. **"
     else
     echo "** FAIL ** Testing failed $?- Exiting release script"
@@ -18,7 +26,7 @@ if python setup.py check --strict
 fi
 
 # check long_description RST formatting
-if python setup.py check --restructuredtext --strict
+if python setup.py check --restructuredtext -s
     then echo "** long_description RST Testing passed. **"
     else
     echo "** FAIL ** Testing failed $?- Exiting release script"
@@ -35,8 +43,8 @@ fi
 
 # check docs build locally
 cd docs
-make clean
-if make html
+$MAKE clean
+if $MAKE html
     then echo "** Doc Build passed. **"
     else
     cd ..
@@ -44,7 +52,15 @@ if make html
     exit 4
 fi
 cd ..
+# check bumpversion
+if bumpversion $1 --verbose --commit --tag patch
+    then echo "** Version Bump passed. **"
+    else
+    echo "** FAIL ** Bumpversion failed - Exiting release script"
+    exit 5
+fi
 # clean up the build directories
+rm -rf dist/
 rm -rf pysftp.egg-info/
 # build
 python setup.py sdist
